requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0 
tortoise==0.1.1
tortoise-orm==0.20.1